added google oauth

+++ app/api/Tickets/[id]/route.js
@@ -1,14 +1,36 @@
 import Ticket from "@/app/models/Ticket";
 import { NextResponse } from "next/server";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 
 export async function GET(request, { params }) {
-  const { id } = params;
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.json(
+      { message: "Unauthorized access" },
+      { status: 401 }
+    );
+  }
 
-  const foundTicket = await Ticket.findOne({ _id: id });
-  return NextResponse.json({ foundTicket }, { status: 200 });
+  const { id } = params;
+  try {
+    const foundTicket = await Ticket.findOne({ _id: id });
+    return NextResponse.json({ foundTicket }, { status: 200 });
+  } catch (error) {
+    console.error(error);
+    return NextResponse.json({ message: "Error", error }, { status: 500 });
+  }
 }
 
 export async function PUT(req, { params }) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.json(
+      { message: "Unauthorized access" },
+      { status: 401 }
+    );
+  }
+
   try {
     const { id } = params;
 
@@ -16,13 +38,15 @@ export async function PUT(req, { params }) {
     const ticketData = body.formData;
     const existingTicket = await Ticket.findById(id);
 
-    const updateTicketData = await Ticket.findByIdAndUpdate(id, {
-      ...ticketData,
-    });
     if (existingTicket.status !== "not started") {
       const { title, description, category, priority } = ticketData;
 
-      if (title !== existingTicket.title || description !== existingTicket.description || category !== existingTicket.category || priority !== existingTicket.priority) {
+      if (
+        title !== existingTicket.title ||
+        description !== existingTicket.description ||
+        category !== existingTicket.category ||
+        priority !== existingTicket.priority
+      ) {
         return NextResponse.json(
           { message: "Cannot edit fields once the status is not 'not started'" },
           { status: 400 }
@@ -30,21 +54,30 @@ export async function PUT(req, { params }) {
       }
     }
 
+    await Ticket.findByIdAndUpdate(id, { ...ticketData });
     return NextResponse.json({ message: "Ticket updated" }, { status: 200 });
   } catch (error) {
-    console.log(error);
+    console.error(error);
     return NextResponse.json({ message: "Error", error }, { status: 500 });
   }
 }
 
 export async function DELETE(req, { params }) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.json(
+      { message: "Unauthorized access" },
+      { status: 401 }
+    );
+  }
+
   try {
     const { id } = params;
 
     await Ticket.findByIdAndDelete(id);
     return NextResponse.json({ message: "Ticket Deleted" }, { status: 200 });
   } catch (error) {
-    console.log(error);
+    console.error(error);
     return NextResponse.json({ message: "Error", error }, { status: 500 });
   }
 }
